search-api/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

coverage
pylint==v2.13.8
pylint-flask
pytest
pytest-cov
pytest-env
pytest-dotenv
pytest-mock
requests
flake8==4.0.1
flake8-blind-except
flake8-debugger
flake8-docstrings
flake8-isort
flake8-quotes
pep8-naming
autopep8
pydocstyle
black==25.1.0
FreezeGun
lovely-pytest-docker
Faker